Preparing to test the single box job.

+++ include/program/jobs/Compute_Path_Bounds.h
@@ -29,17 +29,17 @@
 /// of some other Morse set.
 /// The maximal path lengths, computed with graph algorithms, are also provided.
 template < class Conley_Morse_Graph >
-void Compute_Path_Bounds ( std::map < typename Conley_Morse_Graph::Edge , long > * path_bounds ,
+void Compute_Path_Bounds ( std::map < typename Conley_Morse_Graph::Edge , size_t > * path_bounds ,
   const Conley_Morse_Graph & conley_morse_graph ,
   std::map < typename Conley_Morse_Graph::Vertex , Conley_Morse_Graph const * > & original_cmg ,
   std::map < typename Conley_Morse_Graph::Vertex , typename Conley_Morse_Graph::Vertex > & original_set ,
   std::map < std::pair < Conley_Morse_Graph const * , typename Conley_Morse_Graph::Vertex > , Conley_Morse_Graph const * > & finer_cmg ,
   std::map < Conley_Morse_Graph const * , Conley_Morse_Graph const * > & coarser_cmg ,
   std::map < Conley_Morse_Graph const * , typename Conley_Morse_Graph::Vertex > & coarser_set ,
-  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Vertex , long > > & exit_path_bounds ,
-  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Vertex , long > > & entrance_path_bounds ,
-  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Edge , long > > & path_bounds ,
-  std::map < Conley_Morse_Graph const * , long > & through_path_bound );
+  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Vertex , size_t > > & exit_path_bounds ,
+  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Vertex , size_t > > & entrance_path_bounds ,
+  std::map < Conley_Morse_Graph const * , std::map < typename Conley_Morse_Graph::Edge , size_t > > & path_bounds ,
+  std::map < Conley_Morse_Graph const * , size_t > & through_path_bound );
 
 
 #ifndef _DO_NOT_INCLUDE_HPP_
